Feat: 특정 컬렉션의 모든 핀 조회 api 구현 완료

+++ src/main/java/com/pintogether/backend/dto/ShowPlaceResponseDTO.java
@@ -13,7 +13,7 @@ public class ShowPlaceResponseDTO {
 
     private String roadNameAddress;
 
-    private Long pinCnt;
+    private int pinCnt;
 
     private double latitude;
